+++ src/index.php
@@ -1,7 +1,10 @@
 <?php
 require 'SuQLLog.php';
 require 'TuringMachine.php';
+require 'SuQLReservedWords.php';
 require 'SuQLEntityHelper.php';
+require 'SQLBaseModifier.php';
+require 'SQLModifier.php';
 require 'SuQLHandler.php';
 require 'SQLBuilder.php';
 require 'SuQL.php';